Animals_Classification/requirements.txt
ImageHash==4.3.2
numpy==2.3.3
Pillow==11.3.0
Requests==2.32.5
simple_image_download==0.5
tensorflow==2.17.0
tensorflow_intel==2.17.0
urllib3==2.5.0
